+++ pkg/analysis_server/tool/code_completion/completion_metrics.dart
@@ -10,7 +10,6 @@ import 'package:_fe_analyzer_shared/src/base/syntactic_entity.dart';
 import 'package:analysis_server/src/domains/completion/available_suggestions.dart';
 import 'package:analysis_server/src/protocol/protocol_internal.dart';
 import 'package:analysis_server/src/protocol_server.dart' as protocol;
-import 'package:analysis_server/src/services/completion/completion_core.dart';
 import 'package:analysis_server/src/services/completion/completion_performance.dart';
 import 'package:analysis_server/src/services/completion/dart/completion_manager.dart';
 import 'package:analysis_server/src/services/completion/dart/documentation_cache.dart';
@@ -718,7 +717,7 @@ class CompletionMetricsComputer {
   }
 
   int forEachExpectedCompletion(
-      CompletionRequestImpl request,
+      DartCompletionRequest request,
       MetricsSuggestionListener listener,
       ExpectedCompletion expectedCompletion,
       String? completionLocation,
@@ -1200,19 +1199,11 @@ class CompletionMetricsComputer {
   Future<List<protocol.CompletionSuggestion>> _computeCompletionSuggestions(
       MetricsSuggestionListener listener,
       OperationPerformanceImpl performance,
-      CompletionRequestImpl request,
-      DartdocDirectiveInfo dartdocDirectiveInfo,
-      DocumentationCache? documentationCache,
+      DartCompletionRequest dartRequest,
       [DeclarationsTracker? declarationsTracker,
       protocol.CompletionAvailableSuggestionsParams?
           availableSuggestionsParams]) async {
     List<protocol.CompletionSuggestion> suggestions;
-
-    var dartRequest = DartCompletionRequestImpl.from(
-      request,
-      dartdocDirectiveInfo: dartdocDirectiveInfo,
-      documentationCache: documentationCache,
-    );
 
     if (declarationsTracker == null) {
       // available suggestions == false
@@ -1233,7 +1224,7 @@ class CompletionMetricsComputer {
         listener: listener,
       ).computeSuggestions(dartRequest, performance);
 
-      computeIncludedSetList(declarationsTracker, request.result,
+      computeIncludedSetList(declarationsTracker, dartRequest.result,
           includedSuggestionSetList, includedElementNames);
 
       var includedSuggestionSetMap = {
@@ -1375,27 +1366,21 @@ class CompletionMetricsComputer {
             {required MetricsSuggestionListener listener,
             required CompletionMetrics metrics}) async {
           var stopwatch = Stopwatch()..start();
-          var request = CompletionRequestImpl(
-            resolvedUnitResult,
-            expectedCompletion.offset,
-            CompletionPerformance(),
+          var request = DartCompletionRequest.from(
+            resolvedUnit: resolvedUnitResult,
+            offset: expectedCompletion.offset,
+            documentationCache: documentationCache,
           );
 
           late OpType opType;
           late List<protocol.CompletionSuggestion> suggestions;
-          await request.performance.runRequestOperation(
+          await CompletionPerformance().runRequestOperation(
             (performance) async {
-              var dartRequest = DartCompletionRequestImpl.from(
-                request,
-                documentationCache: documentationCache,
-              );
-              opType = OpType.forCompletion(dartRequest.target, request.offset);
+              opType = OpType.forCompletion(request.target, request.offset);
               suggestions = await _computeCompletionSuggestions(
                 listener,
                 performance,
                 request,
-                dartdocDirectiveInfo,
-                documentationCache,
                 metrics.availableSuggestions ? declarationsTracker : null,
                 metrics.availableSuggestions
                     ? availableSuggestionsParams
@@ -1719,7 +1704,7 @@ class CompletionMetricsOptions {
 class CompletionResult {
   final Place place;
 
-  final CompletionRequestImpl? request;
+  final DartCompletionRequest? request;
 
   final SuggestionData actualSuggestion;